Blackrock Reader Update - NPMK
- current version of Blackrock openNSx didn't seem to work properly when it came to running on new data files on the server.
- put old blackrock files into .rar.
- Look at NPMK/Versions.txt for more info about changes.
- Required changes in phyzzy core
 - openNSx now spits out an int16 array, while decimate requires a double. data is converted to a double. (AnalogIn and Strobe)
 - openNEV (Spikes) no longer has "noparse", but "parse" as an argument. removed "noparse" from argument. NEV.Data.Spikes.Timestamps is now .TimeStamp. Changed in all relevant spots.

+++ preprocessAnalogIn.m
@@ -29,7 +29,7 @@ for i = 1:length(analogInChannels)
   analogInChannelMap(i) = find(analogInHeader.ChannelID == analogInChannels(i));
 end
 analogInChannelMap = analogInChannelMap(analogInChannelMap > 0);
-analogInData = analogInDataRaw(analogInChannelMap,:);
+analogInData = double(analogInDataRaw(analogInChannelMap,:));
 clear analogInDataRaw
 
 % filter the analogIn data using 2 step process.
